+++ ClientApp/package.json
@@ -49,6 +49,7 @@
     "react-scripts": "^5.0.1",
     "react-scroll": "^1.8.3",
     "react-simple-wysiwyg": "^2.0.2",
+    "react-virtuoso": "^2.16.5",
     "reactstrap": "^8.9.0",
     "redux": "4.0.4",
     "redux-thunk": "2.3.0",
